+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add an option to force auto-roll from messages.
 - Add a warning about the public instance ending.
 - Add Portuguese Brazil translation to the public instance warning.